--- a/ui/app/index.html
+++ b/ui/app/index.html
@@ -13,6 +13,7 @@
 
     <link rel="stylesheet" href="{{rootURL}}assets/vendor.css">
     <link rel="stylesheet" href="{{rootURL}}assets/vault.css">
+    <link rel="icon" type="image/png" href="{{rootURL}}favicon.png" />
 
     {{content-for "head-footer"}}
   </head>