--- a/ui/app/styles/components/secret-control-bar.scss
+++ b/ui/app/styles/components/secret-control-bar.scss
@@ -14,10 +14,10 @@
     display: flex;
     flex-direction: column;
     justify-content: center;
-    .switch[type="checkbox"].is-small + label::before {
+    .switch[type='checkbox'].is-small + label::before {
       top: 0.5rem;
     }
-    .switch[type="checkbox"].is-small + label::after {
+    .switch[type='checkbox'].is-small + label::after {
       top: 0.6rem;
     }
   }