--- a/ui/app/styles/components/upgrade-overlay.scss
+++ b/ui/app/styles/components/upgrade-overlay.scss
@@ -11,8 +11,7 @@
   }
 
   .modal-background {
-    background-image: url("/ui/vault-hex.svg"),
-      linear-gradient(90deg, #191a1c, #1b212d);
+    background-image: url('/ui/vault-hex.svg'), linear-gradient(90deg, #191a1c, #1b212d);
     opacity: 0.97;
   }