--- a/ui/app/styles/components/vault-loading.scss
+++ b/ui/app/styles/components/vault-loading.scss
@@ -18,19 +18,19 @@
   }
 
   .vault-loading-order-1 {
-    animation-delay: .1s;
+    animation-delay: 0.1s;
   }
 
   .vault-loading-order-2 {
-    animation-delay: .2s;
+    animation-delay: 0.2s;
   }
 
   .vault-loading-order-3 {
-    animation-delay: .3s;
+    animation-delay: 0.3s;
   }
 
   .vault-loading-order-4 {
-    animation-delay: .4s;
+    animation-delay: 0.4s;
   }
 }