--- a/ui/app/styles/core/breadcrumb.scss
+++ b/ui/app/styles/core/breadcrumb.scss
@@ -47,7 +47,7 @@
 
     &::before {
       color: $blue;
-      content: "❮";
+      content: '❮';
       font-size: 1rem;
       line-height: 1;
       opacity: 0.33;