--- a/ui/app/styles/core/bulma-radio-checkboxes.scss
+++ b/ui/app/styles/core/bulma-radio-checkboxes.scss
@@ -1,17 +1,19 @@
-@import "../sass-svg-uri/svg-uri";
+@import '../sass-svg-uri/svg-uri';
 
-.b-checkbox input[type="checkbox"]:checked + label::before {
+.b-checkbox input[type='checkbox']:checked + label::before {
   border-color: $blue;
 }
 
-.b-checkbox input[type="checkbox"]:checked + label::after,
-.b-checkbox input[type="radio"]:checked + label::after {
+.b-checkbox input[type='checkbox']:checked + label::after,
+.b-checkbox input[type='radio']:checked + label::after {
   font-family: $family-monospace;
   /*checkmark from ionicons*/
-  content: svg-uri('<svg xmlns="http://www.w3.org/2000/svg" width="12" height="12" viewBox="0 0 512 512" fill="#{$blue}"><path d="M461.6 109.6l-54.9-43.3c-1.7-1.4-3.8-2.4-6.2-2.4-2.4 0-4.6 1-6.3 2.5L194.5 323s-78.5-75.5-80.7-77.7c-2.2-2.2-5.1-5.9-9.5-5.9s-6.4 3.1-8.7 5.4c-1.7 1.8-29.7 31.2-43.5 45.8-.8.9-1.3 1.4-2 2.1-1.2 1.7-2 3.6-2 5.7 0 2.2.8 4 2 5.7l2.8 2.6s139.3 133.8 141.6 136.1c2.3 2.3 5.1 5.2 9.2 5.2 4 0 7.3-4.3 9.2-6.2l249.1-320c1.2-1.7 2-3.6 2-5.8 0-2.5-1-4.6-2.4-6.4z"/></svg>');
+  content: svg-uri(
+    '<svg xmlns="http://www.w3.org/2000/svg" width="12" height="12" viewBox="0 0 512 512" fill="#{$blue}"><path d="M461.6 109.6l-54.9-43.3c-1.7-1.4-3.8-2.4-6.2-2.4-2.4 0-4.6 1-6.3 2.5L194.5 323s-78.5-75.5-80.7-77.7c-2.2-2.2-5.1-5.9-9.5-5.9s-6.4 3.1-8.7 5.4c-1.7 1.8-29.7 31.2-43.5 45.8-.8.9-1.3 1.4-2 2.1-1.2 1.7-2 3.6-2 5.7 0 2.2.8 4 2 5.7l2.8 2.6s139.3 133.8 141.6 136.1c2.3 2.3 5.1 5.2 9.2 5.2 4 0 7.3-4.3 9.2-6.2l249.1-320c1.2-1.7 2-3.6 2-5.8 0-2.5-1-4.6-2.4-6.4z"/></svg>'
+  );
 }
 
-.b-checkbox.no-label input[type="checkbox"] + label {
+.b-checkbox.no-label input[type='checkbox'] + label {
   position: absolute;
   top: 0;
 }