--- a/ui/app/styles/core/lists.scss
+++ b/ui/app/styles/core/lists.scss
@@ -4,7 +4,7 @@
   &:before {
     font-size: $size-5;
     color: $white-ter;
-    content: "|";
+    content: '|';
     position: relative;
   }
 }