--- a/ui/app/styles/core/tables.scss
+++ b/ui/app/styles/core/tables.scss
@@ -29,7 +29,7 @@
     border-style: solid;
   }
   tr:hover {
-    background: rgba($grey-lighter, .4);
+    background: rgba($grey-lighter, 0.4);
   }
 }
 .table tbody tr:last-child td,