--- a/ui/app/styles/utils/_colors.scss
+++ b/ui/app/styles/utils/_colors.scss
@@ -1,16 +1,16 @@
 // Colors that will eventually be in Structure package
 
 // Vault Gray
-$vault-gray-010: #FDFDFE;
-$vault-gray-050: #F7FAFC;
-$vault-gray-100: #EBEEF5;
-$vault-gray-200: #D3DBE6;
-$vault-gray-300: #B4BDCC;
-$vault-gray-400: #949FB0;
-$vault-gray-500: #7C8797;
-$vault-gray-600: #5A6370;
-$vault-gray-700: #3C434D;
-$vault-gray-800: #1D2126;
+$vault-gray-010: #fdfdfe;
+$vault-gray-050: #f7fafc;
+$vault-gray-100: #ebeef5;
+$vault-gray-200: #d3dbe6;
+$vault-gray-300: #b4bdcc;
+$vault-gray-400: #949fb0;
+$vault-gray-500: #7c8797;
+$vault-gray-600: #5a6370;
+$vault-gray-700: #3c434d;
+$vault-gray-800: #1d2126;
 $vault-gray-900: #060708;
 $vault-gray: $vault-gray-500;
 $vault-gray-dark: $vault-gray-700;
@@ -23,8 +23,8 @@ $ui-gray-200: #DCE0E6;
 $ui-gray-300: #BAC1CC;
 $ui-gray-500: #6F7682;
 $ui-gray-700: #525761;
-$ui-gray-800: #373A42;
-$ui-gray-900: #1F2124;
+$ui-gray-800: #373a42;
+$ui-gray-900: #1f2124;
 $grey-lightest: $ui-gray-050;
 $grey-light: $ui-gray-300;
 $grey: $ui-gray-500;
@@ -32,13 +32,13 @@ $grey-dark: $ui-gray-700;
 $grey-darkest: $ui-gray-900;
 
 // Blue
-$blue-010: #FBFCFF;
-$blue-050: #F0F5FF;
-$blue-100: #BFD4FF;
-$blue-300: #5B92FF;
+$blue-010: #fbfcff;
+$blue-050: #f0f5ff;
+$blue-100: #bfd4ff;
+$blue-300: #5b92ff;
 $blue-500: #1563ff;
-$blue-700: #0E40A3;
-$blue-900: #061B46;
+$blue-700: #0e40a3;
+$blue-900: #061b46;
 $blue-lightest: $blue-050;
 $blue-light: $blue-300;
 $blue: $blue-500;
@@ -46,13 +46,13 @@ $blue-dark: $blue-700;
 $blue-darkest: $blue-900;
 
 // Red
-$red-010: #FDFAFB;
-$red-050: #F9ECEE;
-$red-100: #EFC7CC;
-$red-300: #DB7D88;
-$red-500: #C73445;
-$red-700: #7F222C;
-$red-900: #370F13;
+$red-010: #fdfafb;
+$red-050: #f9ecee;
+$red-100: #efc7cc;
+$red-300: #db7d88;
+$red-500: #c73445;
+$red-700: #7f222c;
+$red-900: #370f13;
 $red-lightest: $red-050;
 $red-light: $red-300;
 $red: $red-500;
@@ -60,13 +60,13 @@ $red-dark: $red-700;
 $red-darkest: $red-900;
 
 // Green
-$green-010: #FAFDFA;
-$green-050: #ECF7ED;
-$green-100: #C6E9C9;
-$green-300: #7ACC81;
-$green-500: #2EB039;
-$green-700: #1E7125;
-$green-900: #0D3010;
+$green-010: #fafdfa;
+$green-050: #ecf7ed;
+$green-100: #c6e9c9;
+$green-300: #7acc81;
+$green-500: #2eb039;
+$green-700: #1e7125;
+$green-900: #0d3010;
 $green-lightest: $green-050;
 $green-light: $green-300;
 $green: $green-500;
@@ -74,13 +74,13 @@ $green-dark: $green-700;
 $green-darkest: $green-900;
 
 // Orange
-$orange-010: #FFFCFA;
-$orange-050: #FEF4EC;
-$orange-100: #FDE0C8;
-$orange-300: #FBB77F;
-$orange-500: #FA8F37;
-$orange-700: #A05C23;
-$orange-900: #45270F;
+$orange-010: #fffcfa;
+$orange-050: #fef4ec;
+$orange-100: #fde0c8;
+$orange-300: #fbb77f;
+$orange-500: #fa8f37;
+$orange-700: #a05c23;
+$orange-900: #45270f;
 $orange-lightest: $orange-050;
 $orange-light: $orange-300;
 $orange: $orange-500;
@@ -88,12 +88,12 @@ $orange-dark: $orange-700;
 $orange-darkest: $orange-900;
 
 // Yellow
-$yellow-010: #FFFDF6;
-$yellow-050: #FFFBED;
-$yellow-100: #FDEEBA;
-$yellow-300: #FBD95E;
-$yellow-500: #FAC402;
-$yellow-700: #A07D02;
+$yellow-010: #fffdf6;
+$yellow-050: #fffbed;
+$yellow-100: #fdeeba;
+$yellow-300: #fbd95e;
+$yellow-500: #fac402;
+$yellow-700: #a07d02;
 $yellow-900: #453601;
 $yellow-lightest: $yellow-050;
 $yellow-light: $yellow-300;