--- a/ui/ember-cli-build.js
+++ b/ui/ember-cli-build.js
@@ -28,7 +28,7 @@ module.exports = function(defaults) {
     sassOptions: {
       sourceMap: false,
       onlyIncluded: true,
-      implementation: require('sass'),
+      implementation: require('node-sass'),
     },
     autoprefixer: {
       enabled: isProd,