--- a/ui/package.json
+++ b/ui/package.json
@@ -61,7 +61,6 @@
     "ember-cli-content-security-policy": "^1.0.0",
     "ember-cli-dependency-checker": "^3.0.0",
     "ember-cli-eslint": "^4.2.3",
-    "ember-cli-favicon": "~1.0.0",
     "ember-cli-flash": "^1.6.6",
     "ember-cli-htmlbars": "^3.0.0",
     "ember-cli-htmlbars-inline-precompile": "^1.0.3",