--- a/ui/package.json
+++ b/ui/package.json
@@ -105,7 +105,6 @@
     "prettier": "^1.14.3",
     "prettier-eslint-cli": "^4.7.1",
     "qunit-dom": "^0.7.1",
-    "sass": "^1.14.3",
     "sass-svg-uri": "^1.0.0",
     "string.prototype.endswith": "^0.2.0",
     "string.prototype.startswith": "^0.2.0",