When doing development, the build times have crept up over various additions, updates, and development cycles. This removes some things we don't need in development from the build. This also updates `ember-cli` to 3.5.x which has a new version of Broccoli that uses the OS tmp dir and has other performance improvements.

Additionally, this removes ember-cli-favicon which fixed a bug where the linked favicon were being re-loaded on every page navigation in the application.


Here are some unscientific numbers illustrating the before and after - I did each of these a couple of times on the before I took the best time, the after I took the worst one.
<details>
<summary>Output of the ember dev server</summary>
Cold start:
Before:

```
Build successful (49212ms) – Serving on http://localhost:4200/ui/

Slowest Nodes (totalTime => 5% )              | Total (avg)
----------------------------------------------+---------------------
broccoli-persistent-filter:EslintValid... (2) | 9993ms (4996 ms)
ember-auto-import-analyzer (4)                | 6630ms (1657 ms)
Babel: vault (2)                              | 4975ms (2487 ms)
Package /assets/vendor.js (1)                 | 3763ms
```
after:

```
Build successful (15382ms) – Serving on http://localhost:4200/ui/

Slowest Nodes (totalTime => 5% )              | Total (avg)
----------------------------------------------+---------------------
SassCompiler (1)                              | 4317ms
ember-auto-import-analyzer (2)                | 2004ms (1002 ms)
Babel: vault (1)                              | 1829ms
Rollup (3)                                    | 1436ms (478 ms)
Bundler (1)                                   | 1318ms
```

Template file change:
Before:
```
Build successful (5360ms) – Serving on http://localhost:4200/ui/

Slowest Nodes (totalTime => 5% )              | Total (avg)
----------------------------------------------+---------------------
SassCompiler (1)                              | 1510ms
Packaged Application Javascript (1)           | 892ms
broccoli-persistent-filter:Autoprefixe... (1) | 426ms
ember-auto-import-analyzer (4)                | 332ms (83 ms)
Pod Templates (1)                             | 292ms
```

After:
```
Build successful (1004ms) – Serving on http://localhost:4200/ui/

Slowest Nodes (totalTime => 5% )              | Total (avg)
----------------------------------------------+---------------------
Pod Templates (1)                             | 207ms
Processed Application and Dependencies (2)    | 116ms (58 ms)
Babel: vault (1)                              | 96ms
Packaged Application Javascript (1)           | 94ms
Package /assets/vendor.js (1)                 | 74ms
broccoli-persistent-filter:TemplateCom... (6) | 64ms (10 ms)
ember-auto-import-analyzer (2)                | 61ms (30 ms)
```

JS file change:

Before:

```
Build successful (6092ms) – Serving on http://localhost:4200/ui/

Slowest Nodes (totalTime => 5% )              | Total (avg)
----------------------------------------------+---------------------
SassCompiler (1)                              | 1494ms
Packaged Application Javascript (1)           | 755ms
broccoli-persistent-filter:Autoprefixe... (1) | 398ms
Pod Templates (1)                             | 364ms
Concat: App Tests (1)                         | 343ms
SimpleConcatConcat (3)                        | 325ms (108 ms)
```

After:
```
Build successful (1189ms) – Serving on http://localhost:4200/ui/

Slowest Nodes (totalTime => 5% )              | Total (avg)
----------------------------------------------+---------------------
Packaged Application Javascript (1)           | 300ms
Pod Templates (1)                             | 189ms
Package /assets/vendor.js (1)                 | 165ms
Processed Application and Dependencies (2)    | 134ms (67 ms)
ember-auto-import-analyzer (2)                | 63ms (31 ms)
```

SCSS file change:
Before: 
```
Build successful (4683ms) – Serving on http://localhost:4200/ui/

Slowest Nodes (totalTime => 5% )              | Total (avg)
----------------------------------------------+---------------------
SassCompiler (1)                              | 1137ms
Processed Application and Dependencies (2)    | 449ms (224 ms)
broccoli-persistent-filter:Autoprefixe... (1) | 404ms
SimpleConcatConcat (3)                        | 395ms (131 ms)
Pod Templates (1)                             | 313ms
TemplateLinter (1)                            | 256ms
```

After:
```
Build successful (4203ms) – Serving on http://localhost:4200/ui/

Slowest Nodes (totalTime => 5% )              | Total (avg)
----------------------------------------------+---------------------
SassCompiler (1)                              | 3353ms
Pod Templates (1)                             | 216ms
```

After with node-sass:

```
Build successful (1502ms) – Serving on http://localhost:4200/ui/

Slowest Nodes (totalTime => 5% )              | Total (avg)
----------------------------------------------+---------------------
SassCompiler (1)                              | 708ms
Pod Templates (1)                             | 179ms
Package /assets/vendor.js (1)                 | 95ms
Processed Application and Dependencies (2)    | 87ms (43 ms)

file changed styles/components/auth-form.scss
```
</details>

I'll leave some notes on the commit about what we're cutting out to get the speed boost.